+++ lib/core/utils/functions/check_user_auth_token.dart
@@ -3,6 +3,7 @@ import 'package:algonaid_mobile_app/core/routes/paths_routes.dart';
 import 'package:algonaid_mobile_app/core/utils/hive/token_storage.dart';
 import 'package:algonaid_mobile_app/core/utils/cache/shared_pref.dart';
 import 'package:algonaid_mobile_app/core/constants/app_constants.dart';
+import 'package:algonaid_mobile_app/core/network/check_internet.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:jwt_decoder/jwt_decoder.dart';
@@ -44,9 +45,13 @@ Future<void> checkUserAuth([BuildContext? buildContext]) async {
   }
 
   if (isExpired) {
-    await TokenStorage.deleteToken();
-    router.go(Routes.guestHome);
-    return;
+    if (await hasNoInternet()) {
+      debugPrint('checkUserAuth: Token expired, but no internet. Maintaining session for offline access.');
+    } else {
+      await TokenStorage.deleteToken();
+      router.go(Routes.guestHome);
+      return;
+    }
   }
 
   router.go(Routes.homePage);
